actually made the vs solution work

+++ src/enumcs.cpp
@@ -16,14 +16,14 @@ INTERACTION_FLAG operator&(INTERACTION_FLAG lhs, INTERACTION_FLAG rhs) {
 
 EFFECT_RETURN_FLAG operator|(EFFECT_RETURN_FLAG lhs, EFFECT_RETURN_FLAG rhs) {
     return static_cast<EFFECT_RETURN_FLAG>(
-        static_cast<__uint128_t>(lhs) |
-        static_cast<__uint128_t>(rhs)
+        static_cast<Uint8>(lhs) |
+        static_cast<Uint8>(rhs)
     );
 }
 
 EFFECT_RETURN_FLAG operator&(EFFECT_RETURN_FLAG lhs, EFFECT_RETURN_FLAG rhs) {
     return static_cast<EFFECT_RETURN_FLAG>(
-        static_cast<__uint128_t>(lhs) &
-        static_cast<__uint128_t>(rhs)
+        static_cast<Uint8>(lhs) &
+        static_cast<Uint8>(rhs)
     );
 }
